Large rework
Changed many things to be more in-line with the Multiplayer Shooter project.

The default gamemode is now gamemodebase: I believe this is used because GameModeBase lacks multiplayer? Anyway, now I can seperate multiplayer functionality of GOGameMode from the MainMenu.

Large rework of how Widgets are created and added to viewport, many new blueprint macros and such in the GameInstance. This is not without some errors, specifically SplashScreen is giving off errors when it is added to viewport and when removed.

Next up to work on: Network travel errors need to be relinked in the GOGameInstance and fixed. Also, I need to figure out why the server is not spawning in the players, I can achieve this by comparing my logic to the MultiplayerShooter project's logic.