Getting GameMode/PlayerController to spawn characters-- still not functional
It was not working because playercontroller/playerstate beginplay was firing AFTER the onpostlogin event, so the SelectedCharacter variable was not null.

Need to figure out how to store the selected character from the character select screen, I currently am getting it from Game Instance which seems to work in PIE but not in the packaged game, I'd assume because the gamemode does not have access to a player's gameinstance, since gameinstance is not replicated???

Also: Fixed packaging project settings that were not cooking the maps.